+++ app/views/submissions/steps/trial/_step04.html.haml
@@ -1,30 +1,26 @@
-.uploaded-layout{ class: "#{'hidden' unless @content.layout_upload}" }
+.uploaded-trait-scores{ class: "#{'hidden' unless @content.upload}" }
+  %pre.parser-logs.alert.alert-warning.hidden
+  %pre.parser-errors.alert.alert-danger.hidden
+
+  %pre.parser-summary.alert.alert-success{ class: "#{'hidden' unless @content.upload}" }
+    - if @content.upload
+      = raw @content.upload.parser_summary.join('</br>')
+
   .form-group
-    = label_tag :layout, 'Trial layout'
+    = label_tag :upload, 'Uploaded file'
+    %div
+      %span.file-name= @content.upload.try(:file_file_name)
+      = delete_submission_upload_button(@content.upload,
+        class: 'btn btn-default delete-trait-scores-upload')
 
-    = link_to(@content.layout_upload.try(:file_url), target: "_blank") do
-      %img.layout-image{ src: @content.layout_upload.try(:file_url, :small) }
+  .parser-note.alert.alert-info
+    %p
+      Please review the above parsing logs. If there were any critical errors during the process,
+      the red section will try to give some hints regarding their nature. You may then try to fix
+      the file accordingly, delete the current upload and re-upload the fixed file.
 
-    %div.alert.alert-warning
-      %p
-        Uploaded file name:
-        = succeed '.' do
-          %strong
-            %span.file-name= @content.layout_upload.try(:file_file_name)
-      %p
-        Please click the thumbnail above to see the full size version. In order to save the image in BIP,
-        you need to click either
-        %strong Save & edit later
-        or
-        %strong Send
-        button. Use the
-        %strong Delete
-        button right below to replace the image with another one.
-      %p
-        = delete_submission_upload_button(@content.layout_upload,
-          class: 'btn btn-default delete-layout-upload')
+    %p
+      Otherwise, if you think the parsing went properly, use the <b>NEXT</b> button to confirm
+      these changes and proceed with further steps of submission.
 
-= content.hidden_field :layout_upload_id
-
-= render partial: "submissions/annotations", locals: { model: PlantTrial, collapsible: false, form: content }
-= render partial: "submissions/steps/visibility", locals: { form: content }
+= content.hidden_field :upload_id
