+++ app/views/submissions/steps/trial/_step06.html.haml
@@ -1,0 +1,2 @@
+= render partial: "submissions/annotations", locals: { model: PlantTrial, collapsible: false, form: content }
+= render partial: "submissions/steps/visibility", locals: { form: content }
